scripts/requirements.txt
apache-airflow>=2.7.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
